roysched and discount service and basic controller

+++ src/main/java/com/sprint/BookPartnerApplication/repository/RoyschedRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoyschedRepository extends JpaRepository<Roysched, Long> {
+public interface RoyschedRepository extends JpaRepository<Roysched, Integer> {
 }
